0.96 Update Added usage[1] and value[1]
Usage and Value have the possibility to have more than one, and in the last few version of my program it ignored if there was more than one. This is a major issue because it removes spawns for items. Today, I patch this, I have added pseudo variables to hold them until I can figure out how to add two of the same element to an xml object without it overwriting. I think it may be possible to remove name, set attribute, then rename attribute to avoid overwrite.